Pythia

Description

This is the entry point to the library. All services are exported here. See the modules themselves for more documentation.

Since: 0.1

Services

Each service is self-contained in that it should have everything needed for typical usage. Generally, Pythia.Services.X exports everything of interest in Pythia.Services.X.*. That said, there is some additional functionality in Pythia.Data.* and Pythia.Control.Exception that may occasionally be of interest (e.g. optics).

module Pythia.Services.Battery

module Pythia.Services.GlobalIp

module Pythia.Services.Memory

module Pythia.Services.NetInterface

module Pythia.Services.Time

Exceptions

Pythia's error handling is defined in terms of safe-exceptions. Pythia.Control.Exception defines general exceptions that can be thrown from IO; Additionally, services can also throw specific exceptions (e.g. AcpiException from Pythia.Services.Battery.Acpi).

module Pythia.Control.Exception