Pythia.Control.Exception

Description

Provides various exception types.

Since: 0.1

Exception Hierarchy Root

data PythiaException

All specific exceptions thrown by pythia are subtypes of PythiaException.

Since: 0.1

Constructors

forall e.Exception e => MkPythiaException e

toExceptionViaPythia :: Exception e => e -> SomeException

toException via PythiaException. Used for defining an exception as a subtype of PythiaException.

Since: 0.1

fromExceptionViaPythia :: Exception e => SomeException -> Maybe e

fromException via PythiaException. Used for defining an exception as a subtype of PythiaException.

Since: 0.1

Miscellaneous Exceptions

data CommandException

Exceptions encountered while running a shell command.

Examples

Expand

>>> putStrLn $ displayException $ MkCommandException "some command" "an error message"
Command exception. Command: <some command>. Error: <an error message>

Since: 0.1

Constructors

MkCommandException

Fields

command :: Command

The command that was run.

Since: 0.1

errMsg :: Text

The error message received.

Since: 0.1

newtype SomeExceptions

Collects 1 or more exceptions.

Examples

Expand

>>> let ex1 = toException $ MkCommandException "some command" "an error message"
>>> let ex2 = toException $ MkNotSupportedException "app1"
>>> putStrLn $ displayException $ MkSomeExceptions $ ex1 :| [ex2]
Found 2 exception(s)
- Command exception. Command: <some command>. Error: <an error message>
- App not supported: <app1>

Since: 0.1

Constructors

MkSomeExceptions

Fields

unSomeExceptions :: NonEmpty SomeException

Since: 0.1

newtype NotSupportedException

Error for when the current app is not supported.

Examples

Expand

>>> putStrLn $ displayException $ MkNotSupportedException "app1"
App not supported: <app1>

Since: 0.1

Constructors

MkNotSupportedException

Fields

unNotSupportedException :: Text

Since: 0.1

data NoActionsRunException

Error for when no actions are run.

Examples

Expand

>>> putStrLn $ displayException MkNoActionsRunException
No actions run

Since: 0.1

Constructors

MkNoActionsRunException

Optics

_MkPythiaException :: forall e. Exception e => Review PythiaException e

Since: 0.1

_MkSomeExceptions :: Iso' SomeExceptions (NonEmpty SomeException)

Since: 0.1

_MkNotSupportedException :: Iso' NotSupportedException Text

Since: 0.1

_MkNoActionsRunException :: Iso' NoActionsRunException ()

Since: 0.1