Pythia.Data.RunApp

Description

This module provides the RunApp type.

Since: 0.1

Documentation

data RunApp a

Used in conjunction for our services for indicating how we want to run queries. Many is the identity for its monoid instance.

Examples

Expand

>>> Many <> Single "r"
Single "r"

>>> Single "l" <> Many
Single "l"

>>> mempty @(RunApp String)
Many

Since: 0.1

Constructors

Many
Do not specify an application. Attempt to run all, returning the first success.

Since: 0.1

Single a
Runs a single query based on the parameter app.

Since: 0.1

_Many :: forall a. Prism' (RunApp a) ()

_Single :: forall a a. Prism (RunApp a) (RunApp a) a a

Since: 0.1