Pythia.Internal.ShellApp

Description

This module provides the functionality for running shell commands and parsing the result.

Since: 0.1

SimpleShell

data SimpleShell err result

Type for running a "simple" shell command given by Command. The $sel:parser:MkSimpleShell is used to parse the result.

Since: 0.1

Constructors

MkSimpleShell

Fields

command :: Command

The shell command to run.

Since: 0.1

parser :: Text -> Either err result

The parser for the result of running the command.

Since: 0.1

runSimple :: Exception err => SimpleShell err result -> IO result

Runs a simple shell.

Throws:

err: if running the command throws SomeException or a parse error is encountered.

Since: 0.1

Trying Multiple IO

data AppAction m r

Represents some IO app to retrieve a result r. Includes a string name and support query for determining if this action is supported by the current system. Intended for when we want to try multiple actions i.e. tryAppActions.

Since: 0.1

Constructors

MkAppAction

Fields

action :: m r

Since: 0.1

supported :: m Bool

Since: 0.1

name :: Text

Since: 0.1

tryAppActions :: [AppAction IO result] -> IO result

Queries for information via multiple apps. Returns the first success. If any errors are encountered or no actions are run (either because the list is empty or none are supported), an exception is thrown.

Throws:

NoActionsRunException: if no actions are run (i.e. the list is empty or none are supported).

SomeExceptions: if at least one command is run yet there were no successes.

Since: 0.1

tryIOs :: [IO result] -> IO result

Generalized tryAppActions to any IO. Has the same semantics (i.e. returns the first success or throws an exception if none succeeds) without checking for "support".

Throws:

NoActionsRunException: if no actions are run (i.e. the list is empty).

SomeExceptions: if at least one command is run yet there were no successes.

Since: 0.1

Utilities

runCommand :: Command -> IO Text

Runs a Command and returns either the text result or error encountered. This is used by SimpleShell to run its command before the result is parsed. This function is exported for convenience.

Throws:

CommandException: if running the command returns ExitFailure exit code.

Since: 0.1