Pythia.Services.Battery

Description

This module exports battery related services.

Since: 0.1

Queries

queryBattery :: BatteryConfig -> IO Battery

Queries the battery based on the configuration. If $sel:app:MkBatteryConfig is Many then we try supported apps in the following order:

[BatteryAppSysFs, BatteryAppAcpi, BatteryAppUPower]

Throws:

AcpiParseError

SysFsDirNotFound

SysFsBatteryDirNotFound

SysFsFileNotFound

SysFsBatteryParseError

UPowerParseError

CommandException

Since: 0.1

Types

data Battery

Full battery state, including percentage and status data.

Since: 0.1

Constructors

MkBattery

Fields

percentage :: Percentage

Since: 0.1

status :: BatteryStatus

Since: 0.1

newtype Percentage

Represents a percentage.

Since: 0.1

Constructors

MkPercentage

Fields

unPercentage :: LRInterval 0 100 Word8

Since: 0.1

data BatteryStatus

Represents battery charging status.

Since: 0.1

Constructors

Charging
Since: 0.1

Discharging
Since: 0.1

Full
Since: 0.1

Pending
Since: 0.1

Configuration

newtype BatteryConfig

Battery configuration.

>>> mempty @BatteryConfig
MkBatteryConfig {app = Many}

Since: 0.1

Constructors

MkBatteryConfig

Fields

app :: RunApp BatteryApp

Since: 0.1

data BatteryApp

Determines how we should query the system for battery state information.

Since: 0.1

Constructors

BatteryAppSysFs
Uses the sysfs interface i.e. /sys.

Since: 0.1

BatteryAppAcpi
Uses the ACPI utility.

Since: 0.1

BatteryAppUPower
Uses the UPower utility.

Since: 0.1

data RunApp a

Used in conjunction for our services for indicating how we want to run queries. Many is the identity for its monoid instance.

Examples

Expand

>>> Many <> Single "r"
Single "r"

>>> Single "l" <> Many
Single "l"

>>> mempty @(RunApp String)
Many

Since: 0.1

Constructors

Many
Do not specify an application. Attempt to run all, returning the first success.

Since: 0.1

Single a
Runs a single query based on the parameter app.

Since: 0.1

Errors

newtype AcpiParseError

Error parsing acpi output.

Examples

Expand

>>> displayException $ MkAcpiParseError "parse error"
"Acpi parse error: parse error"

Since: 0.1

Constructors

MkAcpiParseError

Fields

unAcpiParseError :: Text

Since: 0.1

data SysFsDirNotFound

Sysfs dir not found error.

Examples

Expand

>>> displayException MkSysFsDirNotFound
"Could not find either sysfs dirs: /sys/class/power_supply, /sysfs/class/power_supply"

Since: 0.1

Constructors

MkSysFsDirNotFound

data SysFsBatteryDirNotFound

Sysfs battery dir not found.

Examples

Expand

>>> displayException MkSysFsBatteryDirNotFound
"Could not find BAT[0-5]? subdirectory under /sys(fs)/class/power_supply"

Since: 0.1

Constructors

MkSysFsBatteryDirNotFound

newtype SysFsFileNotFound

Sysfs file not found.

Examples

Expand

>>> displayException $ MkSysFsFileNotFound "foo"
"Could not find sysfs file: foo"

Since: 0.1

Constructors

MkSysFsFileNotFound

Fields

unSysFsFileNotFound :: Text

Since: 0.1

newtype SysFsBatteryParseError

Sysfs battery parse error.

Examples

Expand

>>> displayException $ MkSysFsBatteryParseError "bad"
"SysFs parse error: bad"

Since: 0.1

Constructors

MkSysFsBatteryParseError

Fields

unSysFsBatteryParseError :: Text

Since: 0.1

data UPowerParseError

Errors that can occur with upower.

Examples

Expand

>>> displayException $ UPowerParseErrorPercentage "output"
"No percentage found in upower output: output"

>>> displayException $ UPowerParseErrorStatus "output"
"No status found in upower output: output"

Since: 0.1

Constructors

UPowerParseErrorPercentage Text
Did not find percentage.

Since: 0.1

UPowerParseErrorStatus Text
Did not find status.

Since: 0.1

Optics

Types

_Charging :: Prism' BatteryStatus ()

_Discharging :: Prism' BatteryStatus ()

_Full :: Prism' BatteryStatus ()

_Pending :: Prism' BatteryStatus ()

Since: 0.1

Configuration

_MkBatteryConfig :: Iso' BatteryConfig (RunApp BatteryApp)

Since: 0.1

_BatteryAppAcpi :: Prism' BatteryApp ()

_BatteryAppSysFs :: Prism' BatteryApp ()

_BatteryAppUPower :: Prism' BatteryApp ()

Since: 0.1

Errors

_MkAcpiParseError :: Iso' AcpiParseError Text

Since: 0.1

_MkSysFsFileNotFound :: Iso' SysFsFileNotFound Text

Since: 0.1

_MkSysFsBatteryParseError :: Iso' SysFsBatteryParseError Text

Since: 0.1

_UPowerParseErrorPercentage :: Prism' UPowerParseError Text

_UPowerParseErrorStatus :: Prism' UPowerParseError Text

Since: 0.1