Pythia.Services.GlobalIp

Description

This modules exports everything needed for retrieving global IP addresses.

Since: 0.1

Queries

queryGlobalIp :: GlobalIpBothConfig -> IO (IpAddress 'Ipv4, IpAddress 'Ipv6)

Queries for IPv4 and IPv6 global IP address based on the configuration. If $sel:app:MkGlobalIpConfig is Many then we try supported apps in the following order:

[GlobalIpAppDig, GlobalIpAppCurl]

Throws:

CommandException

Since: 0.1

queryGlobalIpv4 :: GlobalIpv4Config -> IO (IpAddress 'Ipv4)

queryGlobalIp restricted to IPv4 address only.

Throws:

PythiaException: if an error is encountered (e.g. running a command or parse error).

Since: 0.1

queryGlobalIpv6 :: GlobalIpv6Config -> IO (IpAddress 'Ipv6)

queryGlobalIp restricted to IPv6 address only.

Throws:

PythiaException: if an error is encountered (e.g. running a command or parse error).

Since: 0.1

Types

data IpType

IP types.

Since: 0.1

Constructors

Ipv4
Since: 0.1

Ipv6
Since: 0.1

newtype IpAddress a

Type for an IP address. The type family IpRefinement refines the underlying Text according to the spec.

Ipv4: All characters are digits or dots, and the length is 0 < l < 16.

Ipv6: All characters are hex digits or colons, and the length is 0 < l < 40.

Since: 0.1

Constructors

MkIpAddress

Fields

unIpAddress :: Refined (IpRefinement a) Text

Since: 0.1

Configuration

type GlobalIpv4Config = GlobalIpConfig [UrlSource 'Ipv4]

Type alias for Ipv4 GlobalIpConfig.

Since: 0.1.0.0

type GlobalIpv6Config = GlobalIpConfig [UrlSource 'Ipv6]

Type alias for Ipv6 GlobalIpConfig.

Since: 0.1.0.0

type GlobalIpBothConfig = GlobalIpConfig ([UrlSource 'Ipv4], [UrlSource 'Ipv6])

Type alias for Ipv4 and Ipv6 GlobalIpConfig.

Since: 0.1.0.0

data GlobalIpConfig a

Complete configuration for querying global IP addresses. The Monoid instance will construct a config that tries all apps and has no extra sources.

>>> mempty @(GlobalIpConfig [UrlSource Ipv4])
MkGlobalIpConfig {app = Many, sources = []}

Since: 0.1

Constructors

MkGlobalIpConfig

Fields

app :: RunApp GlobalIpApp

Determines how we want to query.

Since: 0.1

sources :: a

Extra lookup sources. This will be either a single [UrlSource a] or a pair ([UrlSource Ipv4], [UrlSource Ipv6]), depending on which address we want to retrieve.

Since: 0.1

data GlobalIpApp

This type determines what program we use to lookup the ip address.

Since: 0.1

Constructors

GlobalIpAppDig
Uses the dig command to perform a DNS lookup. This is generally the fastest and most reliable.

Since: 0.1

GlobalIpAppCurl
Uses curl to lookup the ip addresses.

Since: 0.1

newtype UrlSource a

Additional URL source for retrieving IP information. The intended app should not be included (i.e. curl or dig), but any desired flags should be.

Examples

Expand

(dig): "resolver1.opendns.com myip.opendns.com"
(dig): -4 TXT o-o.myaddr.l.google.com ns1.google.com
(curl): "http://whatismyip.akamai.com/"
...

Since: 0.1

Constructors

MkUrlSource

Fields

unUrlSource :: Text

Since: 0.1

data RunApp a

Used in conjunction for our services for indicating how we want to run queries. Many is the identity for its monoid instance.

Examples

Expand

>>> Many <> Single "r"
Single "r"

>>> Single "l" <> Many
Single "l"

>>> mempty @(RunApp String)
Many

Since: 0.1

Constructors

Many
Do not specify an application. Attempt to run all, returning the first success.

Since: 0.1

Single a
Runs a single query based on the parameter app.

Since: 0.1