Pythia.Services.Memory

Description

This module exports memory related services.

Since: 0.1

Queries

queryMemory :: MemoryConfig -> IO SystemMemory

Queries the memory based on the configuration. If $sel:app:MkMemoryConfig is Many then we try supported apps in the following order:

[MemoryFree]

Throws:

FreeParseError

CommandException

Since: 0.1

Functions

freeMemory :: SystemMemory -> Memory NonNegative

Returns the amount of free memory.

Since: 0.1

percentageUsed :: SystemMemory -> Percentage

Returns the used memory as a percentage.

Since: 0.1

percentageFree :: SystemMemory -> Percentage

Returns the free memory as a percentage.

Since: 0.1

Types

newtype Memory (f :: Type -> Type)

Represents the current memory usage. The type parameter is some wrapper around the memory intended to enforce an invariant e.g. non-negative.

Since: 0.1

Constructors

MkMemory

Fields

unMemory :: Bytes 'B (f Double)

Since: 0.1

data SystemMemory

Represents the current memory usage.

Since: 0.1

Constructors

MkSystemMemory

Fields

total :: !(Memory Positive)

The total memory on this system.

Since: 0.1

used :: !(Memory NonNegative)

The memory currently in use. This does not include the cache.

Since: 0.1

Configuration

newtype MemoryConfig

Memory configuration.

>>> mempty @MemoryConfig
MkMemoryConfig {app = Many}

Since: 0.1

Constructors

MkMemoryConfig

Fields

app :: RunApp MemoryApp

Since: 0.1

data MemoryApp

Determines how we should query the system for memory usage.

Since: 0.1

Constructors

MemoryAppFree
Uses the free utility.

Since: 0.1

data RunApp a

Used in conjunction for our services for indicating how we want to run queries. Many is the identity for its monoid instance.

Examples

Expand

>>> Many <> Single "r"
Single "r"

>>> Single "l" <> Many
Single "l"

>>> mempty @(RunApp String)
Many

Since: 0.1

Constructors

Many
Do not specify an application. Attempt to run all, returning the first success.

Since: 0.1

Single a
Runs a single query based on the parameter app.

Since: 0.1

Errors

newtype FreeParseError

Errors that can occur with acpi.

Examples

Expand

>>> displayException $ MkFreeParseError "parse error"
"Could not parse memory from: parse error"

Since: 0.1

Constructors

MkFreeParseError

Fields

unFreeParseError :: Text

Since: 0.1

Optics

Types

_MkMemory :: forall f f. Iso (Memory f) (Memory f) (Bytes 'B (f Double)) (Bytes 'B (f Double))

Since: 0.1

Configuration

_MkMemoryConfig :: Iso' MemoryConfig (RunApp MemoryApp)

Since: 0.1

Errors

_MkFreeParseError :: Iso' FreeParseError Text

Since: 0.1