Pythia.Services.Memory.Types

Description

This module provides the core types describing the memory.

Since: 0.1

Documentation

data MemoryApp

Determines how we should query the system for memory usage.

Since: 0.1

Constructors

MemoryAppFree
Uses the free utility.

Since: 0.1

newtype MemoryConfig

Memory configuration.

>>> mempty @MemoryConfig
MkMemoryConfig {app = Many}

Since: 0.1

Constructors

MkMemoryConfig

Fields

app :: RunApp MemoryApp

Since: 0.1

_MkMemoryConfig :: Iso' MemoryConfig (RunApp MemoryApp)

Since: 0.1

newtype Memory (f :: Type -> Type)

Represents the current memory usage. The type parameter is some wrapper around the memory intended to enforce an invariant e.g. non-negative.

Since: 0.1

Constructors

MkMemory

Fields

unMemory :: Bytes 'B (f Double)

Since: 0.1

_MkMemory :: forall f f. Iso (Memory f) (Memory f) (Bytes 'B (f Double)) (Bytes 'B (f Double))

Since: 0.1

data SystemMemory

Represents the current memory usage.

Since: 0.1

Constructors

MkSystemMemory

Fields

total :: !(Memory Positive)

The total memory on this system.

Since: 0.1

used :: !(Memory NonNegative)

The memory currently in use. This does not include the cache.

Since: 0.1