Pythia.Services.NetInterface

Description

This module exports interface related services.

Since: 0.1

Queries

queryNetInterfaces :: NetInterfaceConfig -> IO NetInterfaces

Queries for all network interface data. If the NetInterfaceConfig's app is Many then we try all NetInterfaceApps supported by this system, in the following order:

[NetInterfaceAppNmCli, NetInterfaceAppIp]

Throws:

PythiaException: if an error is encountered (e.g. running a command or parse error).

Since: 0.1

queryNetInterface :: Device -> NetInterfaceConfig -> IO NetInterface

Like queryNetInterfaces but returns data for a single device.

Throws:

DeviceNotFound

IpParseError

NmCliParseError

CommandException

Since: 0.1

Functions

findUp :: NetInterfaces -> Maybe NetInterface

Takes the first NetInterface that has state NetStateUp, according to NetInterfaceState's Ord:

Ethernet < Wifi < Wifi_P2P < Loopback < Tun

Examples

>>> findUp $ MkNetInterfaces []
Nothing

>>> :{
  let wifiUp = MkNetInterface "" (Just Wifi) NetStateUp (Just "WifiUp") mempty mempty
      wifiNetStateDown = MkNetInterface "" (Just Wifi) NetStateDown (Just "WifiNetStateDown") mempty mempty
      loopUp = MkNetInterface "" (Just Loopback) NetStateUp (Just "LoopUp") mempty mempty
   in findUp $ MkNetInterfaces [loopUp, wifiNetStateDown, wifiUp]
:}
Just (MkNetInterface {device = MkDevice {unDevice = ""}, ntype = Just Wifi, state = NetStateUp, name = Just "WifiUp", ipv4s = MkIpAddresses {unIpAddresses = []}, ipv6s = MkIpAddresses {unIpAddresses = []}})

Since: 0.1

Types

newtype NetInterfaces

Since: 0.1

Constructors

MkNetInterfaces

Fields

unNetInterfaces :: [NetInterface]

Since: 0.1

data NetInterface

Full connection data.

Since: 0.1

Constructors

MkNetInterface

Fields

device :: Device

Since: 0.1

ntype :: Maybe NetInterfaceType

Since: 0.1

state :: NetInterfaceState

Since: 0.1

name :: Maybe Text

The name of the connection (e.g. Wifi SSID).

Since: 0.1

ipv4s :: IpAddresses 'Ipv4

Since: 0.1

ipv6s :: IpAddresses 'Ipv6

Since: 0.1

data NetInterfaceState

Various connection states.

Since: 0.1

Constructors

NetStateUp
Since: 0.1

NetStateDown
Since: 0.1

NetStateUnknown Text
Since: 0.1

data NetInterfaceType

Various connection types.

Since: 0.1

Constructors

Ethernet
Since: 0.1

Wifi
Since: 0.1

Wifi_P2P
Since: 0.1

Loopback
Since: 0.1

Tun
Since: 0.1

newtype Device

Newtype wrapper over a network device name.

Since: 0.1

Constructors

MkDevice

Fields

unDevice :: Text

Since: 0.1

data IpType

IP types.

Since: 0.1

Constructors

Ipv4
Since: 0.1

Ipv6
Since: 0.1

newtype IpAddress a

Type for an IP address. The type family IpRefinement refines the underlying Text according to the spec.

Ipv4: All characters are digits or dots, and the length is 0 < l < 16.

Ipv6: All characters are hex digits or colons, and the length is 0 < l < 40.

Since: 0.1

Constructors

MkIpAddress

Fields

unIpAddress :: Refined (IpRefinement a) Text

Since: 0.1

Configuration

newtype NetInterfaceConfig

Complete configuration for querying network interfaces.

>>> mempty @NetInterfaceConfig
MkNetInterfaceConfig {app = Many}

Since: 0.1

Constructors

MkNetInterfaceConfig

Fields

app :: RunApp NetInterfaceApp

Since: 0.1

data NetInterfaceApp

Determines how we should query the system for interface state information.

Since: 0.1

Constructors

NetInterfaceAppNmCli
Uses the Network Manager cli utility.

Since: 0.1

NetInterfaceAppIp
Uses the 'ip' utility.

Since: 0.1

data RunApp a

Used in conjunction for our services for indicating how we want to run queries. Many is the identity for its monoid instance.

Examples

Expand

>>> Many <> Single "r"
Single "r"

>>> Single "l" <> Many
Single "l"

>>> mempty @(RunApp String)
Many

Since: 0.1

Constructors

Many
Do not specify an application. Attempt to run all, returning the first success.

Since: 0.1

Single a
Runs a single query based on the parameter app.

Since: 0.1

Errors

newtype DeviceNotFound

Exception for when we cannot find a desired device.

Examples

Expand

>>> displayException $ MkDeviceNotFound "bad device"
"Device not found: bad device"

Since: 0.1

Constructors

MkDeviceNotFound

Fields

unDeviceNotFound :: Device

Since: 0.1

newtype IpParseError

Error parsing ip output.

Examples

Expand

>>> displayException $ MkIpParseError "parse error"
"Ip parse error: parse error"

Since: 0.1

Constructors

MkIpParseError

Fields

unIpParseError :: Text

Since: 0.1

newtype NmCliParseError

Error parsing nmcli output.

Examples

Expand

>>> displayException $ MkNmCliParseError "parse error"
"NmCli parse error: parse error"

Since: 0.1

Constructors

MkNmCliParseError

Fields

unNmCliParseError :: Text

Since: 0.1

Optics

Types

_MkNetInterfaces :: Iso' NetInterfaces [NetInterface]

Since: 0.1

_NetStateUp :: Prism' NetInterfaceState ()

_NetStateDown :: Prism' NetInterfaceState ()

_NetStateUnknown :: Prism' NetInterfaceState Text

Since: 0.1

_Ethernet :: Prism' NetInterfaceType ()

_Wifi :: Prism' NetInterfaceType ()

_Wifi_P2P :: Prism' NetInterfaceType ()

_Loopback :: Prism' NetInterfaceType ()

_Tun :: Prism' NetInterfaceType ()

Since: 0.1

_MkDevice :: Iso' Device Text

Since: 0.1

_Ipv4 :: Prism' IpType ()

_Ipv6 :: Prism' IpType ()

Since: 0.1

_MkIpAddress :: forall a a. Iso (IpAddress a) (IpAddress a) (Refined (IpRefinement a) Text) (Refined (IpRefinement a) Text)

Since: 0.1

Configuration

_MkNetInterfaceConfig :: Iso' NetInterfaceConfig (RunApp NetInterfaceApp)

Since: 0.1

_NetInterfaceAppNmCli :: Prism' NetInterfaceApp ()

_NetInterfaceAppIp :: Prism' NetInterfaceApp ()

Since: 0.1

Errors

_MkDeviceNotFound :: Iso' DeviceNotFound Device

Since: 0.1

_MkIpParseError :: Iso' IpParseError Text

Since: 0.1

_MkNmCliParseError :: Iso' NmCliParseError Text

Since: 0.1