Pythia.Services.NetInterface.Ip

Description

This module provides functionality for retrieving network connection information using ip utility.

Since: 0.1

Query

netInterfaceShellApp :: (MonadPathReader m, MonadThrow m, MonadTypedProcess m) => m NetInterfaces

Ip query for NetInterface.

Since: 0.1

supported :: MonadPathReader m => m Bool

Returns a boolean determining if this program is supported on the current system.

Since: 0.1

Misc

newtype IpParseError

Error parsing ip output.

Examples

Expand

>>> displayException $ MkIpParseError "parse error"
"Ip parse error: parse error"

Since: 0.1

Constructors

MkIpParseError Text

parseInterfaces :: Text -> Either IpParseError NetInterfaces

Attempts to parse the output of IP.

Since: 0.1