Pythia.Services.NetInterface.NmCli

Description

This module provides functionality for retrieving network connection information using nmcli.

Since: 0.1

Query

netInterfaceShellApp :: IO NetInterfaces

NmCli query for NetInterfaces.

Throws:

NmCliException: if something goes wrong (i.e. exception while running the command, or we have a parse error).

Since: 0.1

supported :: IO Bool

Returns a boolean determining if this program is supported on the current system.

Since: 0.1

Misc

newtype NmCliParseError

Error parsing nmcli output.

Examples

Expand

>>> displayException $ MkNmCliParseError "parse error"
"NmCli parse error: parse error"

Since: 0.1

Constructors

MkNmCliParseError

Fields

unNmCliParseError :: Text

Since: 0.1

_MkNmCliParseError :: Iso' NmCliParseError Text

Since: 0.1

parseInterfaces :: Text -> Either NmCliParseError NetInterfaces

Attemps to parse the output of nmcli.

Since: 0.1