Pythia.Services.NetInterface.Types

Description

Provides network interface types.

Since: 0.1

Configuration

data NetInterfaceApp

Determines how we should query the system for interface state information.

Since: 0.1

Constructors

NetInterfaceAppNmCli
Uses the Network Manager cli utility.

Since: 0.1

NetInterfaceAppIp
Uses the 'ip' utility.

Since: 0.1

NetInterface Fields

data NetInterfaceType

Various connection types.

Since: 0.1

Constructors

Ethernet
Since: 0.1

Wifi
Since: 0.1

Wifi_P2P
Since: 0.1

Loopback
Since: 0.1

Tun
Since: 0.1

data NetInterfaceState

Various connection states.

Since: 0.1

Constructors

NetStateUp
Since: 0.1

NetStateDown
Since: 0.1

NetStateUnknown Text
Since: 0.1

data NetInterface

Full connection data.

Since: 0.1

Constructors

MkNetInterface

Fields

device :: Device

Since: 0.1

ntype :: Maybe NetInterfaceType

Since: 0.1

state :: NetInterfaceState

Since: 0.1

name :: Maybe Text

The name of the connection (e.g. Wifi SSID).

Since: 0.1

ipv4s :: IpAddresses Ipv4

Since: 0.1

ipv6s :: IpAddresses Ipv6

Since: 0.1

newtype NetInterfaces

Since: 0.1

Constructors

MkNetInterfaces

Fields

unNetInterfaces :: [NetInterface]

Errors

newtype DeviceNotFound

Exception for when we cannot find a desired device.

Examples

Expand

>>> displayException $ MkDeviceNotFound "bad device"
"Device not found: bad device"

Since: 0.1

Constructors

MkDeviceNotFound Device

Optics

_NetInterfaceAppNmCli :: Prism' NetInterfaceApp ()

Since: 0.1

_NetInterfaceAppIp :: Prism' NetInterfaceApp ()

Since: 0.1

_Ethernet :: Prism' NetInterfaceType ()

Since: 0.1

_Wifi :: Prism' NetInterfaceType ()

Since: 0.1

_Wifi_P2P :: Prism' NetInterfaceType ()

Since: 0.1

_Loopback :: Prism' NetInterfaceType ()

Since: 0.1

_Tun :: Prism' NetInterfaceType ()

Since: 0.1

_NetStateUp :: Prism' NetInterfaceState ()

Since: 0.1

_NetStateDown :: Prism' NetInterfaceState ()

Since: 0.1

_NetStateUnknown :: Prism' NetInterfaceState Text

Since: 0.1