Pythia.Services.Types.Network

Description

Provides common network types.

Since: 0.1

IP Types

data IpType

IP types.

Since: 0.1

Constructors

Ipv4
Since: 0.1

Ipv6
Since: 0.1

newtype IpAddress a

Type for an IP address. The type family IpRefinement refines the underlying Text according to the spec.

Ipv4: All characters are digits or dots, and the length is 0 < l < 16.

Ipv6: All characters are hex digits or colons, and the length is 0 < l < 40.

Since: 0.1

Constructors

MkIpAddress

Fields

unIpAddress :: Refined (IpRefinement a) Text

Since: 0.1

newtype IpAddresses a

Since: 0.1

Constructors

MkIpAddresses

Fields

unIpAddresses :: [IpAddress a]

Since: 0.1

Refinements

type family IpRefinement a where ...

Maps IpType to its Text refinement.

Since: 0.1

Equations

IpRefinement Ipv4 = Ipv4Refinement

IpRefinement Ipv6 = Ipv6Refinement

data Ipv4Refinement

IPv4 Refinement. We implement a custom type here so we get better error messages. Text must satisfy:

Length in (0, 16).

Chars are all decimal digits or dots.

Examples

Expand

>>> R.refine @Ipv4Refinement @Text "192.168.1.2"
Right (Refined "192.168.1.2")

>>> trim $ R.refine @Ipv4Refinement @Text ""
"Left   The predicate (Ipv4Refinement) failed with the message: Expected IPv4 with length (0, 16). Received empty."

>>> trim $ R.refine @Ipv4Refinement @Text "192.168.111.222.7"
"Left   The predicate (Ipv4Refinement) failed with the message: Expected IPv4 address with length (0, 16). Received '192.168.111.222.7' of length 17"

>>> trim $ R.refine @Ipv4Refinement @Text "192.168x1.2"
"Left   The predicate (Ipv4Refinement) failed with the message: IPv4 address should only contain decimal digits or dots. Received invalid: 192.168x1.2"

Since: 0.1

data Ipv6Refinement

IPv6 Refinement. We implement a custom type here so we get better error messages. Text must satisfy:

Length in (0, 40).

Chars are all hex digits or colons.

Examples

Expand

>>> R.refine @Ipv6Refinement @Text "fe80::a328:482:5263:10b8"
Right (Refined "fe80::a328:482:5263:10b8")

>>> trim $ R.refine @Ipv6Refinement @Text ""
"Left   The predicate (Ipv6Refinement) failed with the message: Expected IPv6 of length (0, 40). Received empty."

>>> trim $ R.refine @Ipv6Refinement @Text "fe80:a328:4822:5263:10b8:4062:10d3:16ac:"
"Left   The predicate (Ipv6Refinement) failed with the message: Expected IPv6 with length (0, 40). Received 'fe80:a328:4822:5263:10b8:4062:10d3:16ac:' of length 40"

>>> trim $ R.refine @Ipv6Refinement @Text "fe80::a328:482:5263:10b8x"
"Left   The predicate (Ipv6Refinement) failed with the message: IPv6 address should only contain hex digits or colons. Received invalid: fe80::a328:482:5263:10b8x"

Since: 0.1

Network Device

newtype Device

Newtype wrapper over a network device name.

Since: 0.1

Constructors

MkDevice

Fields

unDevice :: Text

Since: 0.1

Optics

_Ipv4 :: Prism' IpType ()

Since: 0.1

_Ipv6 :: Prism' IpType ()

Since: 0.1