Pythia.Utils

Description

This module provides common utilities.

Since: 0.1

Folding

foldAlt :: (Foldable t, Alternative f) => (a -> f b) -> t a -> f b

Similar to foldMap but for Alternative.

Examples

Expand

>>> foldAlt (\c -> if even c then Just c else Nothing) [1,2,3,4]
Just 2

>>> foldAlt (\c -> if even c then Just c else Nothing) [1,3]
Nothing

Since: 0.1

mAlt :: Alternative f => Maybe (f a) -> f a

Convenience function for mapping a Maybe to its underlying Alternative.

Examples

Expand

>>> mAlt @[] Nothing
[]

>>> mAlt @[] (Just [1,2,3])
[1,2,3]

Since: 0.1

Parsing

takeLine :: (Ord e, Stream s, Token s ~ Char) => Parsec e s (Tokens s)

takeLineLabel with no label.

Examples

Expand

>>> parseTest @Void takeLine "some text 123 \n"
"some text 123 "

>>> parseTest @Void takeLine "some text 123"
1:14:
  |
1 | some text 123
  |              ^
unexpected end of input
expecting end of line

Since: 0.1

takeLineLabel :: (Ord e, Stream s, Token s ~ Char) => Maybe String -> Parsec e s (Tokens s)

Variant of takeLine taking in a label.

Examples

Expand

>>> parseTest @Void (takeLineLabel (Just "a label")) "some text 123"
1:14:
  |
1 | some text 123
  |              ^
unexpected end of input
expecting a label or end of line

Since: 0.1

takeLine_ :: (Ord e, Stream s, Token s ~ Char) => Parsec e s ()

Takes everything up to the first new line, returns unit.

Examples

Expand

>>> parseTest @Void takeLine_ "some text 123\n"
()

Since: 0.1

exeSupported :: String -> IO Bool

Determines if the executable represented by the string parameter is supported on this system.

Since: 0.1

Pretty Printing

class Pretty a where

Overloaded conversion to Doc.

Laws:

output should be pretty. :-)

Minimal complete definition

pretty

Methods

pretty :: a -> Doc ann

>>> pretty 1 <+> pretty "hello" <+> pretty 1.234
1 hello 1.234

prettyList :: [a] -> Doc ann

prettyList is only used to define the instance Pretty a => Pretty [a]. In normal circumstances only the pretty function is used.

>>> prettyList [1, 23, 456]
[1, 23, 456]

data Doc ann

The abstract data type Doc ann represents pretty documents that have been annotated with data of type ann.

More specifically, a value of type Doc represents a non-empty set of possible layouts of a document. The layout functions select one of these possibilities, taking into account things like the width of the output document.

The annotation is an arbitrary piece of data associated with (part of) a document. Annotations may be used by the rendering backends in order to display output differently, such as

color information (e.g. when rendering to the terminal)

mouseover text (e.g. when rendering to rich HTML)

whether to show something or not (to allow simple or detailed versions)

The simplest way to display a Doc is via the Show class.

>>> putStrLn (show (vsep ["hello", "world"]))
hello
world

(<+>) :: Doc ann -> Doc ann -> Doc ann infixr 6

(x <+> y) concatenates document x and y with a space in between.

>>> "hello" <+> "world"
hello world

x <+> y = x <> space <> y

comma :: Doc ann

>>> comma
,

punctuate

Arguments

:: Doc ann
Punctuation, e.g. comma

-> [Doc ann]

-> [Doc ann]

(punctuate p xs) appends p to all but the last document in xs.

>>> let docs = punctuate comma (Util.words "lorem ipsum dolor sit amet")
>>> putDocW 80 (hsep docs)
lorem, ipsum, dolor, sit, amet

The separators are put at the end of the entries, which we can see if we position the result vertically:

>>> putDocW 20 (vsep docs)
lorem,
ipsum,
dolor,
sit,
amet

If you want put the commas in front of their elements instead of at the end, you should use tupled or, in general, encloseSep.

hsep :: [Doc ann] -> Doc ann

(hsep xs) concatenates all documents xs horizontally with <+>, i.e. it puts a space between all entries.

>>> let docs = Util.words "lorem ipsum dolor sit amet"

>>> hsep docs
lorem ipsum dolor sit amet

hsep does not introduce line breaks on its own, even when the page is too narrow:

>>> putDocW 5 (hsep docs)
lorem ipsum dolor sit amet

For automatic line breaks, consider using fillSep instead.

vsep :: [Doc ann] -> Doc ann

(vsep xs) concatenates all documents xs above each other. If a group undoes the line breaks inserted by vsep, the documents are separated with a space instead.

Using vsep alone yields

>>> "prefix" <+> vsep ["text", "to", "lay", "out"]
prefix text
to
lay
out

grouping a vsep separates the documents with a space if it fits the page (and does nothing otherwise). See the sep convenience function for this use case.

The align function can be used to align the documents under their first element:

>>> "prefix" <+> align (vsep ["text", "to", "lay", "out"])
prefix text
       to
       lay
       out

Since grouping a vsep is rather common, sep is a built-in for doing that.

layoutCompact :: Doc ann1 -> SimpleDocStream ann2

(layoutCompact x) lays out the document x without adding any indentation and without preserving annotations. Since no 'pretty' printing is involved, this layouter is very fast. The resulting output contains fewer characters than a prettyprinted version and can be used for output that is read by other programs.

>>> let doc = hang 4 (vsep ["lorem", "ipsum", hang 4 (vsep ["dolor", "sit"])])
>>> doc
lorem
    ipsum
    dolor
        sit

>>> let putDocCompact = renderIO System.IO.stdout . layoutCompact
>>> putDocCompact doc
lorem
ipsum
dolor
sit

renderStrict :: SimpleDocStream ann -> Text

(renderStrict sdoc) takes the output sdoc from a rendering function and transforms it to strict text.

prettyToText :: Pretty a => a -> Text

Converts a type with a Pretty instance to Text.

Since: 0.1

Miscellaneous

headMaybe :: [a] -> Maybe a

Total version of head.

Examples

Expand

>>> headMaybe []
Nothing

>>> headMaybe [3, 4]
Just 3

Since: 0.1

eitherToBool :: Either a b -> Bool

Maps Left to False, Right to True.

Examples

Expand

>>> eitherToBool (Left ())
False

>>> eitherToBool (Right ())
True

Since: 0.1