Pythia.Internal.ShellApp

Description

This module provides the functionality for running shell commands and parsing the result.

Since: 0.1

SimpleShell

data SimpleShell m err result

Type for running a "simple" shell command given by Command. The $sel:parser:MkSimpleShell is used to parse the result.

Since: 0.1

Constructors

MkSimpleShell

Fields

command :: Command

The shell command to run.

Since: 0.1

isSupported :: m Bool

Determines if the shell command is supported on this system.

Since: 0.1

parser :: Text -> Either err result

The parser for the result of running the command.

Since: 0.1

runSimple :: forall m err result. (Exception err, MonadThrow m, MonadTypedProcess m) => SimpleShell m err result -> m result

Runs a simple shell.

Throws:

NotSupportedException: if the command is not supported on this system.

err: if running the command throws SomeException or a parse error is encountered.

Since: 0.1

Trying Multiple IO

tryIOs :: MonadCatch m => [m result] -> m result

Generalized tryAppActions to any IO. Has the same semantics (i.e. returns the first success or throws an exception if none succeeds) without checking for "support".

Throws:

NoActionsRunException: if no actions are run (i.e. the list is empty).

SomeExceptions: if at least one command is run yet there were no successes.

Since: 0.1

Utilities

runCommand :: (MonadThrow m, MonadTypedProcess m) => Command -> m Text

Runs a Command and returns either the text result or error encountered. This is used by SimpleShell to run its command before the result is parsed. This function is exported for convenience.

Throws:

CommandException: if running the command returns ExitFailure exit code.

Since: 0.1