Pythia.Services.Battery.Types

Description

This module provides the core types describing the battery.

Since: 0.1

Configuration

data BatteryApp

Determines how we should query the system for battery state information.

Since: 0.1

Constructors

BatteryAppSysFs
Uses the sysfs interface i.e. /sys.

Since: 0.1

BatteryAppAcpi
Uses the ACPI utility.

Since: 0.1

BatteryAppUPower
Uses the UPower utility.

Since: 0.1

Battery Fields

data BatteryStatus

Represents battery charging status.

Since: 0.1

Constructors

Charging
Since: 0.1

Discharging
Since: 0.1

Full
Since: 0.1

Pending
Since: 0.1

data Battery

Full battery state, including percentage and status data.

Since: 0.1

Constructors

MkBattery

Fields

percentage :: Percentage

Since: 0.1

status :: BatteryStatus

Since: 0.1

Optics

_BatteryAppAcpi :: Prism' BatteryApp ()

Since: 0.1

_BatteryAppSysFs :: Prism' BatteryApp ()

Since: 0.1

_BatteryAppUPower :: Prism' BatteryApp ()

Since: 0.1

_Charging :: Prism' BatteryStatus ()

Since: 0.1

_Discharging :: Prism' BatteryStatus ()

Since: 0.1

_Full :: Prism' BatteryStatus ()

Since: 0.1

_Pending :: Prism' BatteryStatus ()

Since: 0.1