Pythia.Services.GlobalIp

Description

This modules exports everything needed for retrieving global IP addresses.

Since: 0.1

Queries

queryGlobalIp :: (MonadCatch m, MonadTypedProcess m) => GlobalIpBothConfig -> m (IpAddress Ipv4, IpAddress Ipv6)

Queries for IPv4 and IPv6 global IP address based on the configuration.

Since: 0.1

queryGlobalIpv4 :: (MonadCatch m, MonadTypedProcess m) => GlobalIpv4Config -> m (IpAddress Ipv4)

queryGlobalIp restricted to IPv4 address only.

Since: 0.1

queryGlobalIpv6 :: (MonadCatch m, MonadTypedProcess m) => GlobalIpv6Config -> m (IpAddress Ipv6)

queryGlobalIp restricted to IPv6 address only.

Since: 0.1

Types

data IpType

IP types.

Since: 0.1

Constructors

Ipv4
Since: 0.1

Ipv6
Since: 0.1

newtype IpAddress a

Type for an IP address. The type family IpRefinement refines the underlying Text according to the spec.

Ipv4: All characters are digits or dots, and the length is 0 < l < 16.

Ipv6: All characters are hex digits or colons, and the length is 0 < l < 40.

Since: 0.1

Constructors

MkIpAddress (Refined (IpRefinement a) Text)

Configuration

type GlobalIpv4Config = GlobalIpConfig [UrlSource Ipv4]

Type alias for Ipv4 GlobalIpConfig.

Since: 0.1.0.0

type GlobalIpv6Config = GlobalIpConfig [UrlSource Ipv6]

Type alias for Ipv6 GlobalIpConfig.

Since: 0.1.0.0

type GlobalIpBothConfig = GlobalIpConfig ([UrlSource Ipv4], [UrlSource Ipv6])

Type alias for Ipv4 and Ipv6 GlobalIpConfig.

Since: 0.1.0.0

data GlobalIpConfig a

Complete configuration for querying global IP addresses.

Since: 0.1

Constructors

MkGlobalIpConfig

Fields

app :: GlobalIpApp

Determines how we want to query.

Since: 0.1

sources :: a

Extra lookup sources. This will be either a single [UrlSource a] or a pair ([UrlSource Ipv4], [UrlSource Ipv6]), depending on which address we want to retrieve.

Since: 0.1

data GlobalIpApp

This type determines what program we use to lookup the ip address.

Since: 0.1

Constructors

GlobalIpAppCurl
Uses curl to lookup the ip addresses.

Since: 0.1

GlobalIpAppDig
Uses the dig command to perform a DNS lookup. This is generally the fastest and most reliable.

Since: 0.1

newtype UrlSource a

Additional URL source for retrieving IP information. The intended app should not be included (i.e. curl or dig), but any desired flags should be.

Examples

Expand

(dig): "resolver1.opendns.com myip.opendns.com"
(dig): -4 TXT o-o.myaddr.l.google.com ns1.google.com
(curl): "http://whatismyip.akamai.com/"
...

Since: 0.1

Constructors

MkUrlSource

Fields

unUrlSource :: Text