Pythia.Services.Memory.Types

Description

This module provides the core types describing the memory.

Since: 0.1

Documentation

data MemoryApp

Determines how we should query the system for memory usage.

Since: 0.1

Constructors

MemoryAppFree
Uses the free utility.

Since: 0.1

newtype Memory

Represents the current memory usage. The type parameter is some wrapper around the memory intended to enforce an invariant e.g. non-negative.

Since: 0.1

Constructors

MkMemory

Fields

unMemory :: Bytes B Natural

data SystemMemory

Represents the current memory usage.

Since: 0.1

Constructors

MkSystemMemory

Fields

total :: Memory

The total memory on this system.

Since: 0.1

used :: Memory

The memory currently in use. This does not include the cache.

Since: 0.1