Pythia.Services.NetInterface

Description

This module exports interface related services.

Since: 0.1

Queries

queryNetInterfaces :: (MonadPathReader m, MonadThrow m, MonadTypedProcess m) => NetInterfaceApp -> m NetInterfaces

Queries for all network interface data.

Since: 0.1

queryNetInterface :: (MonadPathReader m, MonadThrow m, MonadTypedProcess m) => Device -> NetInterfaceApp -> m NetInterface

Like queryNetInterfaces but returns data for a single device.

Since: 0.1

Functions

findUp :: NetInterfaces -> Maybe NetInterface

Takes the first NetInterface that has state NetStateUp, according to NetInterfaceState's Ord:

Ethernet < Wifi < Wifi_P2P < Loopback < Tun

Examples

>>> findUp $ MkNetInterfaces []
Nothing

>>> :{
  let wifiUp = MkNetInterface "" (Just Wifi) NetStateUp (Just "WifiUp") mempty mempty
      wifiNetStateDown = MkNetInterface "" (Just Wifi) NetStateDown (Just "WifiNetStateDown") mempty mempty
      loopUp = MkNetInterface "" (Just Loopback) NetStateUp (Just "LoopUp") mempty mempty
   in findUp $ MkNetInterfaces [loopUp, wifiNetStateDown, wifiUp]
:}
Just (MkNetInterface {device = MkDevice {unDevice = ""}, ntype = Just Wifi, state = NetStateUp, name = Just "WifiUp", ipv4s = MkIpAddresses {unIpAddresses = []}, ipv6s = MkIpAddresses {unIpAddresses = []}})

Since: 0.1

Types

newtype NetInterfaces

Since: 0.1

Constructors

MkNetInterfaces

Fields

unNetInterfaces :: [NetInterface]

data NetInterface

Full connection data.

Since: 0.1

Constructors

MkNetInterface

Fields

device :: Device

Since: 0.1

ntype :: Maybe NetInterfaceType

Since: 0.1

state :: NetInterfaceState

Since: 0.1

name :: Maybe Text

The name of the connection (e.g. Wifi SSID).

Since: 0.1

ipv4s :: IpAddresses Ipv4

Since: 0.1

ipv6s :: IpAddresses Ipv6

Since: 0.1

data NetInterfaceState

Various connection states.

Since: 0.1

Constructors

NetStateUp
Since: 0.1

NetStateDown
Since: 0.1

NetStateUnknown Text
Since: 0.1

data NetInterfaceType

Various connection types.

Since: 0.1

Constructors

Ethernet
Since: 0.1

Wifi
Since: 0.1

Wifi_P2P
Since: 0.1

Loopback
Since: 0.1

Tun
Since: 0.1

newtype Device

Newtype wrapper over a network device name.

Since: 0.1

Constructors

MkDevice

Fields

unDevice :: Text

Since: 0.1

data IpType

IP types.

Since: 0.1

Constructors

Ipv4
Since: 0.1

Ipv6
Since: 0.1

newtype IpAddress a

Type for an IP address. The type family IpRefinement refines the underlying Text according to the spec.

Ipv4: All characters are digits or dots, and the length is 0 < l < 16.

Ipv6: All characters are hex digits or colons, and the length is 0 < l < 40.

Since: 0.1

Constructors

MkIpAddress

Fields

unIpAddress :: Refined (IpRefinement a) Text

Since: 0.1

Configuration

data NetInterfaceApp

Determines how we should query the system for interface state information.

Since: 0.1

Constructors

NetInterfaceAppNmCli
Uses the Network Manager cli utility.

Since: 0.1

NetInterfaceAppIp
Uses the 'ip' utility.

Since: 0.1